--- HandCricketApplication/IHandling.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace HandCricketApplication
-{
-    interface IHandling
-    {
-        void PlayingHandle(string battingPerson, string bowlingPerson);
-
-    }
-}
