+++ HandCricketApplication/Repository/IHandling.cs
@@ -1,0 +1,8 @@
+﻿namespace HandCricketApplication.Repository
+{
+      interface IHandling
+    {
+        void PlayingHandle(string battingPerson, string bowlingPerson);
+
+    }
+}
