+++ HandCricketApplication/Repository/IPlayOperation.cs
@@ -1,0 +1,9 @@
+﻿namespace HandCricketApplication.Repository
+{
+      interface IPlayOperation
+    {
+        void SinglePlay(string battingPerson, string bowlingPerson);
+
+        void GroupPlaying(params string[] players);
+    }
+}
